Cast UNet to target dtype after loading LoRA weights
LoRA weights may load in bfloat16 from the checkpoint, causing dtype
mismatches when the base model is float32. Explicitly cast unet to
the target dtype after load_lora_weights().

+++ animatediff_motion_lora_tilt_down/pytorch/loader.py
@@ -92,7 +92,7 @@ class ModelLoader(ForgeModel):
 
         self.pipeline.load_lora_weights(LORA_REPO)
 
-        return self.pipeline.unet
+        return self.pipeline.unet.to(dtype)
 
     def load_inputs(self, **kwargs) -> Any:
         """Prepare synthetic UNet inputs for tilt-down motion inference.
